backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
